videoparser: use official includes for vkvideo_parser
Remove from videoparser.h, the video parser structures and
use instead the same includes as in vk_video_samples.
It will be easier to update the value and avoid to miss
VPS addition in VkParserHevcPictureData for example.
